requirements.txt
fastapi
uvicorn[standard]
httpx
pytest-asyncio
flask_restful
Flask